# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CommonBeacon REST API",
    "version": "1.0.0",
    "description": "Implemented /api/v1 contract. Same-origin cookie sessions, CSRF on every mutation. Roles include MEMBER, MODERATOR, ADMINISTRATOR. Ordinary conversation writes reject archived boards; reporting/moderation remain available. Text is plain text, trimmed as documented and bounded in UTF-16 units (JSON Schema maxLength counts Unicode code points). All page offsets must fit a signed 32-bit integer. Business failures use ProblemDetail; infrastructure failures can have other response shapes. Health probes /api/health and /healthz are deployment endpoints, outside this contract."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/admin/data/imports": {
      "post": {
        "operationId": "createCompanyImport",
        "summary": "Initiate a quarantined native or Discourse company import",
        "tags": [
          "transfers"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "description": "Requires an ACTIVE ADMINISTRATOR, CSRF, UUID Idempotency-Key and IMPORT_UPLOAD grant. Accepts formatVersion=1, recentAuthGrant and optional provider NATIVE (default) or DISCOURSE. Provider is bound into request idempotency. Creates UPLOADING; matching retries return the original job. Uploads expire after one hour. No domain data is changed. All responses use Cache-Control: no-store.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid request",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Sign in required",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden or invalid CSRF/grant/ticket",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version/state/idempotency conflict or download already active",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Request body or reserved transfer capacity exceeds limits",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Transfer storage unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "201": {
            "description": "Upload initiated; original job on matching replay",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferJob"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportUploadRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/data/imports/{id}/archive": {
      "put": {
        "operationId": "uploadCompanyArchive",
        "summary": "Stream an archive into private quarantine",
        "tags": [
          "transfers"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "description": "Requester-only UPLOADING import, ACTIVE ADMINISTRATOR and CSRF required. Raw application/zip for NATIVE (64 MiB) or application/json for DISCOURSE (8 MiB), including chunked bodies. Media type must match the immutable provider. No multipart, filename, path or URL. Ten-minute deadline and 30-second idle timeout. Successful complete upload stores SHA-256 and moves to UPLOADED; a worker automatically inspects it. Interrupted uploads return to UPLOADING and may restart in full; cancelled/expired jobs require new initiation. Another active upload/worker or retry after completion returns 409. Proxy failures may have non-ProblemDetail bodies. No domain data is written.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid request",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Sign in required",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden or invalid CSRF/grant/ticket",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version/state/idempotency conflict or download already active",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Request body or reserved transfer capacity exceeds limits",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Transfer storage unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "200": {
            "description": "Entire archive accepted into quarantine; UPLOADED",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferJob"
                }
              }
            }
          },
          "404": {
            "description": "Job not found or not owned",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Only raw application/zip is supported",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/zip": {
              "schema": {
                "type": "string",
                "format": "binary",
                "maxLength": 67108864
              }
            },
            "application/json": {
              "schema": {
                "type": "string",
                "format": "binary",
                "description": "CommonBeacon Discourse bundle v1; strict source version, shape and count checks."
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/data/imports/{id}/inspection": {
      "get": {
        "operationId": "getArchiveInspection",
        "summary": "Get safe quarantine inspection results",
        "tags": [
          "transfers"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "description": "Requester-only ACTIVE ADMINISTRATOR. Returns 409 until an inspection is available. Results have at most 1000 file/line/code issues and no raw source values. valid means native archive inspection passed, not eligibility to activate. Both valid and rejected archives move to REVIEW_REQUIRED; activationAvailable is always false. Request the separate dry-run endpoint for private staging and target review; confirmation requires a fresh eligible dry-run review.",
        "responses": {
          "400": {
            "description": "Invalid request",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Sign in required",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden or invalid CSRF/grant/ticket",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Job not found or not owned",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Version/state/idempotency conflict or download already active",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Transfer storage unavailable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "200": {
            "description": "Inspection result, never activation approval",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveInspection"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID",
        "description": "HttpOnly SameSite=Lax session cookie. Local HTTP profile disables Secure; production profile enables it. In-memory sessions expire on backend restart."
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ]
      },
...
      "TransferJob": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "COMPANY_EXPORT",
              "PERSONAL_EXPORT",
              "COMPANY_IMPORT"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "READY",
              "FAILED",
              "CANCELLED",
              "UPLOADING",
              "UPLOADED",
              "VALIDATING",
              "REVIEW_REQUIRED",
              "READY_TO_COMMIT",
              "COMMITTING",
              "COMPLETED"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "processedRecords": {
            "type": "integer",
            "minimum": 0
          },
          "errorCode": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              null,
              "WORK_FAILED",
              "AUTHORIZATION_REVOKED",
              "ATTEMPTS_EXHAUSTED",
              "JOB_EXPIRED",
              "ARTIFACT_MISSING",
              "ACTIVATION_RECONCILIATION_REQUIRED",
              "INVALID_SOURCE_DATA",
              "TRANSFER_LIMIT_EXCEEDED",
              "SNAPSHOT_TIMEOUT"
            ]
          },
          "artifactAvailable": {
            "type": "boolean"
          },
          "allowedActions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CANCEL",
                "DOWNLOAD"
              ]
            }
          },
          "provider": {
            "type": "string",
            "enum": [
              "NATIVE",
              "DISCOURSE"
            ],
            "default": "NATIVE",
            "description": "Immutable import provider. Discourse accepts only CommonBeacon bundle v1 from pinned Discourse 3.5.0."
          }
        },
        "required": [
          "id",
          "kind",
          "state",
          "version",
          "createdAt",
          "updatedAt",
          "expiresAt",
          "processedRecords",
          "errorCode",
          "artifactAvailable",
          "allowedActions"
        ],
        "additionalProperties": false
      },
...
      "ImportUploadRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "formatVersion",
          "recentAuthGrant"
        ],
        "properties": {
          "formatVersion": {
            "type": "integer",
            "const": 1
          },
          "recentAuthGrant": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "provider": {
            "type": "string",
            "enum": [
              "NATIVE",
              "DISCOURSE"
            ],
            "default": "NATIVE",
            "description": "Immutable import provider. Discourse accepts only CommonBeacon bundle v1 from pinned Discourse 3.5.0."
          }
        }
      },
      "ArchiveInspection": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "jobId",
          "archiveSha256",
          "valid",
          "rowsChecked",
          "totalErrors",
          "issues",
          "inspectedAt",
          "activationAvailable"
        ],
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "archiveSha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "valid": {
            "type": "boolean"
          },
          "rowsChecked": {
            "type": "integer",
            "minimum": 0,
            "maximum": 40001
          },
          "totalErrors": {
            "type": "integer",
            "minimum": 0
          },
          "issues": {
            "type": "array",
            "maxItems": 1000,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "file",
                "line",
                "code"
              ],
              "properties": {
                "file": {
                  "type": "string",
                  "maxLength": 32
                },
                "line": {
                  "type": "integer",
                  "minimum": 0
                },
                "code": {
                  "type": "string",
                  "maxLength": 64
                }
              }
            }
          },
          "inspectedAt": {
            "type": "string",
            "format": "date-time"
          },
          "activationAvailable": {
            "type": "boolean",
            "const": false
          }
        }
      },
...
    }
  }
}